requirements.txt
azure-functions
backports.entry-points-selectable==1.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
distlib==0.3.2
dominate==2.5.1
feedgen==0.9.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.3
MarkupSafe==1.1.1
pipenv==2021.5.29
platformdirs==2.3.0
pymongo==3.10.1
python-dateutil==2.8.2
requests==2.23.0
six==1.16.0
urllib3==1.25.9
virtualenv==20.7.2
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==0.16.1
